Implement validation; introduce concept of validation levels.
The "validation level" is the degree to which fairport will go to ensure the data is reads from disk (or, the structure you're trying to open on top of that data) is vaild. If the validation level is sufficently high, it'll "fail fast" at time of open and you'll be able to debug your problem. If the validation level is too low, it'll continue on happily, probably crashing at some point in the future.

There are currently three validation levels, FAIRPORT_VALIDATION_LEVEL_NONE, _WEAK, and _FULL. #define the one you want before including any fairport header. See primatives.h for details.

This change (adding validation) also necessitated some other changes in the page interface(s). Some refactoring occured as well.


git-svn-id: https://fairport.svn.codeplex.com/svn/fairport/trunk@37976 c102de6d-fbcd-4082-88ef-c2bab1378b82