Add more static_assert calls for sizes and offsets
These were checked with Microsoft Visual C++ Express 10, and they cover
all the fields that were giving me trouble on the Mac, plus as many of
the others as I could easily calculate fixed sizes for.

Note that we need to move some static assertions to bt_page, because
nbt_nonleaf_page is not a "plain old data" class, and the C++ standard
doesn't actually allow us to use offsetof on it.  GCC raises a warning
when we try, and -Wall therefore causes compilation to fail.