Align disk-level structures using MSVC-compatible layout
On Windows, the ulonglong type (and any structures containing it) will
be aligned to the nearest 8 bytes.  But on Linux and the Mac, they'll be
aligned to the nearest 4 by default, breaking much of the low-level disk
code in pstsdk.

I tried a variety of approaches to fixing this problem, including
'#pragma ms_struct on', but they only approach which worked reliably on
Ubuntu was to add __attributes__((ms_struct)) to each data structure
representing an on-disk data structure.  After some experimentaion, I
decided that the easiest way to do this portably was to define a
PSTSDK_MS_STRUCT macro that conditionally expanded to the appropriate
declaration under GCC.

There are two special cases here.  We add PSTSDK_MS_STRUCT to the 'guid'
structure, even though it aligns correctly by default, as a simple
matter of documentation.  And we omit PSTSDK_MS_STRUCT from a handful of
structures in disk.h that already have their own '#pragma pack(2)'
declarations.