Convert between Windows Unicode and platform-specific wstring
Unfortunately, wchar_t and std::wstring are poorly defined by the C++
standard.  For example, wchar_t may be 2 or 4 bytes wide, and it
contains an unknown encoding (not even necessarily Unicode, or so I've
been told).

We can work around this by writing functions to convert to and from
std::wstring using libiconv, which is readily available on most
non-Windows platforms, and which is typically used for this purpose.

We modify the CMakeLists.txt files to link against iconv (if available).
But we really ought to decide at cmake-time whether or not iconv is
actually necessary, and issue an error message if we need it and can't
find it.  This will need to wait until my CMake skills improve.