Restore most MSVC compiler flags to CMakeLists.txt
The orignal CMakeLists.txt patch removed some MSVC compiler flags that
were present in test/makefile.  This patch restores the following flags
from the orignal list:

  /DPSTSDK_VALIDATION_LEVEL_FULL /EHsc /nologo /W4 /WX

These all seem to be generally applicable, and worth preserving.

However, the following flags were omitted:

  /DEBUG /Od /MDd /Zi

These appear to be specific to Debug mode (with the possible exception
of /Zi), which is normally handled by CMake's support for generating
both a Debug and a Release configuration.