Add a node::get_db() function
When writing code to manipulate node objects, I occasionally need access
to the database's property_bag or name_id_map.  The latter, for example,
is useful when writing a function to dump all the properties of a node.
The former is needed to compute a node's MAPI ID.

One possible solution would be to always pass around both the
shared_db_ptr and a node to any function which might need to access
these properties.  But this would add a rarely-used shared_db_ptr to
many functions which would not otherwise require it.

This patch exposes the m_db member of node as get_db(), making it
accessible to functions which need to do low-level processing on a node.