Compute MAPI PR_ENTRYID for messages and attachments
In e-discovery applications, messages will often be referred to by a
MAPI PR_ENTRYID.  This is a computed property, as described here:

  http://pstsdk.codeplex.com/Thread/View.aspx?ThreadId=215111

Because this is a fairly useful property, and because it's relatively
hard to calculate, I feel it might be worth including somewhere in
pstsdk.

I wanted to implement a node::get_entry_id function, but because
get_entry_id needs access to the message store property bag, this would
have made the ndb layer depend on the ltp layer.

Thus, this patch provides a property_bag::get_entry_id() function, which
does not require any additional dependencies between layers.